growth_factor_RKI_timeseries() added
Added a function for calculating the growth factor (german:
Vervielfaeltigung pro Tag, bestandsspezifischer Wachstumsfaktor) in
RKI data which have been restructured with group_RKI_timeseries(). It
is based on KumAnzahlFall.

The function works on arbitrary groupings made with
group_RKI_timeseries().

How it is implemented:

- 1) Calculate the cartesian product of the restructered data and the
  same data. (cf. self cross join).

- 2) Filter rows where Meldedatum.x == Meldedatum.y + 1 Day

- 3) Calculate the quotient Meldedatum.x / Meldedatum.y

- 4) Replace Inf and NaN values with NA.

- 5) left join to add new column to restructered data.

For performance reasons, an inner join is used instead of a cross join
where possible. The inner join also allows calculation for the growth
factor for arbitrary groupings made with group_RKI_timeseries().